Created Tables of the Models Cardinality Relation between two tables
Product ------------ Category
1 --> 1
Many <-- 1
The relation between Product and Category is Many:1

We made a class a table by using the annotation @Entity and made an attribute a primary key by using @Id annotation.
Also, we inserted a foreign key using the annotation @ManyToOne, essentially the annotation describing the cardinality between the two tables.